finished refator
