+++ src-tauri/Cargo.toml
@@ -27,14 +27,15 @@ tauri-build = { version = "1.0.0-beta-rc.0" }
 cabr2_config = { path = "./cabr2_config/" }
 cabr2_load_save = { path = "./cabr2_load_save/", features = ["beryllium", "cabr2"] }
 cabr2_logger = { path = "./cabr2_logger/" }
-cabr2_search = { path = "./cabr2_search/", features = [ "tauri_plugin", "gestis" ] }
+cabr2_search = { path = "./cabr2_search/", features = ["gestis"] }
 log = "0.4.14"
-tauri = { version = "1.0.0-beta-rc.4", features = ["api-all"] }
+tauri = { version = "1.0.0-beta-rc.4", features =["api-all"], optional = true }
 
 [features]
 custom-protocol = [ "tauri/custom-protocol" ]
 portable = [ "cabr2_config/portable" ]
-default = [ "custom-protocol", "portable" ]
+tauri_app = [ "tauri", "custom-protocol", "cabr2_config/tauri_plugin", "cabr2_load_save/tauri_plugin", "cabr2_search/tauri_plugin" ]
+default = [ "portable", "tauri_app" ]
 
 [workspace]
 members = [
