+++ src-tauri/cabr2_config/Cargo.toml
@@ -17,9 +17,11 @@ lazy_static = "1.4.0"
 log = "0.4.14"
 serde = { version = "1.0.126", features = ["derive"] }
 serde_json = "1.0.64"
-tauri = { version = "1.0.0-beta-rc.4", default-features = false }
+tauri = { version = "1.0.0-beta-rc.4", default-features = false, optional = true }
 thiserror = "1.0.24"
 toml = { version = "0.5.8", features = ["preserve_order"] }
 
 [features]
-portable = [ ]
+portable = []
+tauri_plugin = ["tauri"]
+default = []
