+++ src-tauri/cabr2_config/src/plugin.rs
@@ -1,4 +1,5 @@
 use serde_json::Value;
+use tauri::{plugin::Plugin, InvokeMessage, Params, Window};
 
 use crate::{
   error::Result,
@@ -39,4 +40,38 @@ pub fn get_localized_strings(language: String) -> Result<Value> {
 #[tauri::command]
 pub fn get_prompt_html(name: String) -> Result<String> {
   handler::get_prompt_html(name)
+}
+
+pub struct Config<M: Params> {
+  invoke_handler: Box<dyn Fn(InvokeMessage<M>) + Send + Sync>,
+}
+
+impl<M: Params> std::default::Default for Config<M> {
+  fn default() -> Self {
+    Self {
+      invoke_handler: Box::new(tauri::generate_handler![
+        get_program_version,
+        get_config,
+        save_config,
+        get_hazard_symbols,
+        get_available_languages,
+        get_localized_strings,
+        get_prompt_html,
+      ]),
+    }
+  }
+}
+
+impl<M: Params> Plugin<M> for Config<M> {
+  fn name(&self) -> &'static str {
+    "cabr2_config"
+  }
+
+  fn extend_api(&mut self, message: InvokeMessage<M>) {
+    (self.invoke_handler)(message)
+  }
+
+  fn created(&mut self, _: Window<M>) {
+    log::trace!("plugin created");
+  }
 }
