+++ src-tauri/cabr2_load_save/Cargo.toml
@@ -20,12 +20,14 @@ lopdf = { version = "0.26.0", optional = true }
 quick-xml = { version = "0.22.0", features = ["serialize"], optional = true }
 regex = { version = "1.4.6", optional = true }
 serde = { version = "1.0.126", features = ["derive"] }
-serde_json = "1.0.64" # needed for tauri::command
-tauri = { version = "1.0.0-beta-rc.4", default-features = false }
+serde_json = { version = "1.0.64", optional = true }
+tauri = { version = "1.0.0-beta-rc.4", default-features = false, optional = true }
 thiserror = "1.0.24"
 wkhtmltopdf = { version = "0.4.0", optional = true }
 
 [features]
 beryllium = ["chrono", "quick-xml", "regex"]
-cabr2 = []
-pdf = ["chrono", "handlebars", "lopdf", "wkhtmltopdf"]
+cabr2 = ["serde_json"]
+pdf = ["chrono", "handlebars", "lopdf", "serde_json", "wkhtmltopdf"]
+tauri_plugin = ["tauri", "serde_json"]
+default = []
