+++ src-tauri/cabr2_logger/src/plugin.rs
@@ -1,9 +1,12 @@
 use serde_json::{to_string_pretty, Value};
+use tauri::{plugin::Plugin, InvokeMessage, Params, Window};
 
 use cabr2_types::logging::LogLevel;
 
+use crate::setup_logger;
+
 #[tauri::command]
-pub fn log(level: LogLevel, path: String, messages: Vec<Value>) -> Result<(), String> {
+pub fn log_cmd(level: LogLevel, path: String, messages: Vec<Value>) -> Result<(), String> {
   let mut formatted_messages = Vec::with_capacity(messages.len());
   for message in messages {
     match message {
@@ -25,4 +28,32 @@ pub fn log(level: LogLevel, path: String, messages: Vec<Value>) -> Result<(), St
   }
 
   Ok(())
+}
+
+pub struct Logger<M: Params> {
+  invoke_handler: Box<dyn Fn(InvokeMessage<M>) + Send + Sync>,
+}
+
+impl<M: Params> Logger<M> {
+  pub fn new() -> Self {
+    setup_logger().unwrap();
+
+    Logger {
+      invoke_handler: Box::new(tauri::generate_handler![log_cmd]),
+    }
+  }
+}
+
+impl<M: Params> Plugin<M> for Logger<M> {
+  fn name(&self) -> &'static str {
+    "cabr2_logger"
+  }
+
+  fn extend_api(&mut self, message: InvokeMessage<M>) {
+    (self.invoke_handler)(message)
+  }
+
+  fn created(&mut self, _: Window<M>) {
+    log::trace!("plugin created");
+  }
 }
